task_queue: Make en- and dequeueing thread safe
Introduce a mutex for the task queue, making the enqueue and dequeue
operations thread safe, by locking and unlocking the mutex before and
after the critical sections respectfully.
Also introduce a condition variable signalling to dequeue() that the
queue is not empty, making dequeue a blocking call until it successfully
retrieves a task.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>