task_queue: Return task directly from retrieve call
Instead of writing the retrieved task into a passed in pointer,
rather directly return it.
This was impossible while the task_queue_dequeue_status enum was
still needed to signal success or failure of the dequeue operation.
But as that was removed by the previous commit, the task can now be
directly returned, making for easier semantics when calling retireve.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>