queue: Make head and tail members queue_item_t pointers
Instead of the head and tail members of the queue_t struct being indices into an
items array (which was removed in the previous commit), rather make them
pointers to the queue item structs at the beginning and end of the linked list
of queue items, which will be used as the iternal queue representation.

Also update the queue_init() function to initialize the changed members
accordingly.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>