queue.c: Pass retrieved item's reference to queue_item_destroy()
In queue_destroy() where all of the remaining items in the queue get destroyed
as well, the call to queue_item_destroy() needed to be updated to conform to the
new API of passing in a double pointer.
As no reference of a return value can be taken, store the returned queue_item_t
in a local variable to pass in its address.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>