task/task.h: Create general module internal header
Change the task/task.h header file to be a general module internal header file
that can be included by code inside of the task module/directory. It in turn
includes the internal typedefs and internal function declarations, but also the
external function declarations, which the internal code can also use.

Signed-off-by: Jim Gerth <jim.frederic.gerth@gmail.com>